Expanded the tests.
We now have precompressed data streams with lots of different
compression parameters (all thoses tested for in the round trips).
The test harness now validates it can decompress these streams too.

This grew the test data hugely, so the size of each file has been
considerably reduced.  The original tests, plus all their compressed
variants, will now be put to a new repository called htscodecs-corpus.

This repository, if checked out into the tests subdir, will be
automatically used by the htscodecs test framework.  Thus it could be
a submodule and we choose whether or not to do a recursive clone.
(For now I'll leave this as a manual choice.)