Share bit-packing method between rANS4x16 and arith.
Note this changes the rANS binary output as it was packing in the
opposite endian to arith.  Purely accidental!

Also removed the ALLOW_UAC stuff and just recoded it to use memcpy.
On some compilers it won't generate greate code, but most modern
compilers do a decent job of generating much the same code as the old
unaligned access stuff did, but in a portable fashion.