Bug fixed r4x16 packing, plus optimisations.
The change to bit packing algorithm (to match arith) meant the
unpacked data starts with n > 16, not sym-per-byte == 1.  This also
demonstrated an error in the compress_bounds arithmetic.

Also detected a huge (50-60%) speed difference for r4x16 -o0 decoding
between -O2 and -O3 for both gcc and clang.  Enable O3 by default (can
be overridden still), but also manually unrolled the code.  O3 is
still faster than O2, but the difference is not so much.