rANS4x16 now has frequencies stored summing to power of 2.
For the order-1 case the previous version stored the frequency table
unnormalised if it summed to < TOTFREQ_O1 as this shrinks a
potentially expensive 2D frequency table.  The decoder then has to
normalise these prior to decoding.

Unfortunately this makes the normalisation step part of the
specification, and it's complex to define (and there are likely better
normalisation algorithms than the one I have).

Instead we normalise to a power of 2 prior to writing the frequency
table, although not necessarily the same power of 2 as TOTFREQ_O1.
The decoder still needs to renormalise as each sub-frequency table in
the 2D F array needs to have the same totfreq, but the decoder
renormalisation now is a trivial bit-shift and easy/well defined.

Performance (size/time) is comparable.