Imported JavaScript code from https://github.com/jkbonfield/hts-specs/tree/cram_reference/
This is demonstration code and not intended for production usage.
It is here as a proof of principle for the codec specification
document.

Also fixed a minor bug in the C arith_dynamic that caused
round-tripping via JS to C to fail (but C to C was fine due to a more
relaxed bit-packer).