Fix rANS4x16 RLE in JavaScript.
If no RLE elements are suitable, the C version disables the RLE
request.  That's trickier here as we've already written out the format
byte, so we take a different approach and force at least one symbol to
be RLEd.  This is needed because zero symbols is a synonym for all 256
symbols.