Further speed up to r4x16 O1 decoding.
The "sfb" version with a single struct for sym, freq and base is
optimal when decoding large blocks of data with 10-bit frequencies,
but it has a high memory footprint and takes quite a bit of setup
time.

This means on small blocks or with 12-bit frequency (4x the memory) we
incurr disproportionately slower setup overheads and/or higher cache
miss rates.

So I added back the logic used in the 4x8 decoder with a single
element in the array (sym) and a separate freq/base lookup table
indexed on sym.

This and previous commit means scram_flagstat -b on a NovaSeq CRAM is
now around 8% faster to decode.