fix undefined behaviour (int overflow and neg shift).
This only occurs on malformed data, so the impact is it could fail in a
slightly different manner perhaps.  Valid data cannot trigger the
undefined behaviour as it requires overflows in the frequency tables.

Fixed by explicitly using unsigned data types.  Did not affect
released copies of htslib as this is the first time we've used this
submodule.

Credit to OSS-Fuzz
Fixes oss-fuzz 29802