Add an arbitrary maxname on the name tokeniser.
This is set to about 15Gb.  It's still high, but it prevents the most
egregious excesses.  The purpose is simply to prevent malformed data
from requesting obscene amounts of memory (although in theory that may
still be possible if the caller is also doing multi-threading).

That's equivalent to 10 million records in a block.  1 million may be
more reasonable, but it's also something I've used in extreme cases so
if feels to close to proper usage to limit it at that value.

Detected while looking at OSS-Fuzz 29796, but this isn't that bug
(which has already been fixed elsewhere).