fix undefined shift.
Google fuzzer identified this in tokenise_name3.c.

The rANS_static.c change was found looking for others via grep.  I'm
unsure if it's possible to trigger (although it seems likely) and if
so why it hasn't been found yet, but protecting it too just incase.

Credit to OSS-Fuzz
Fixes oss-fuzz 29956