Fix an overflow bug in the unstripe function.
This was caused by a subtracting froman unsigned value and wrapping
around.  The unstripe function was a static function common to both
the rANS4x16 and arith_dynamic codecs, so it's now been moved to a
utils.h file as a static inline.  This file isn't installed.

It's named "utils.h" as I plan on migrating some other common code to
it too.  Notably the histogram analysis from rANS4x8 and rANS4x16
differ in implementation, but do the same thing.

Credit to OSS-Fuzz
Fixes oss-fuzz 30087