fix access of uninitialise memory in decode for name tokeniser.
If a delta operation, DDELTA, DDELTA0 or MATCH then it needs to
validate that we've not gone beyond the number of tokens used in the
previous name.

This can even trigger on the very first read, so we need to make sure
we also initialise the first name last_ntok too (pnum==cnum==0 for
first name).

Credit to OSS-Fuzz
Fixes oss-fuzz 31210
Fixes oss-fuzz 31215