Fix a bug with name tokeniser and variable number of names per slice.
If the number of names increases by exactly 1 between successive
slices, and the old value was the size of our allocated buffer, then
we don't grow it by one.  This causes the last decoded name to return
-1 instead of 0 as we do the max_names check before doing t0 < 0,
which in turn gives a failure to decode slice error message.

It could be fixed in two ways, by shifting where the end of names
check happens, or fixing the allocation.  I chose the latter as it
would take more head scratching to determine why the allocation has
"++max_names" in it.