Update to release 1.1.1
The only change is updating the HTSCODECS_VERSION macro in htscodecs.h
also.  I forgot to do this for 1.1, so this is a tiny change.  (The
NEWS file just glosses over the existance of 1.1 and jumps straight to
1.1.1.)

My apologies.

Also added a MAINTAINERS.md file as an aide memoir for future releases.