Speed up r4x16pr -o0 encode on gcc11.
Switch to a branchless encoder renorm.  This makes a significant
difference on encode speeds for complex data.  It's slightly
detrimenal on low complexity (eg Illumina 4-qual as-is), but we
usually perform a transform on such data first to convert to high
complexity (RLE and/or PACK), in which case this is still a small
gain.

Enc/Dec times. (Dec code not chaning, but showing reproducability)

gcc11    q40 -o0        q4 -o192
OLD      258.6/564.3    512/1193
NEW      337.2/570.9    534/1177

gcc(7)
OLD      255.3/655.7    489/1192
NEW      287.5/652.7    500/1197

icc21
OLD      279.8/457.5    530/994
NEW      343.5/457.9    554/995

clang10
OLD      256.4/660.7    515/1218
NEW      383/5/658.7    562/1207