Work around issues with Mingw64 having a tiny stack.
The updates to hist1_4 and hist8 use large local buffers.  This fails on
mingw64 unless e.g. -Wl,--stack,4000000 is used.
The alternative is the same strategy employed in rANS elsewhere of
using pthread_once to allocate the local memory as a global pointer,
once per thread in use.

If threads aren't available it reverts to a traditional malloc/free
cycle.  (This was benchmarked and it's generally still better for large
blocks than the older 8-bit frequency counting methods.)