Add a realloc to input buffer for rans4x16pr.
This allows us to stress test fuzzer outputs that crash the decoder
when overflowing the input buffer.  By reallocing it down to the exact
input size the address sanitizer will spot read-overruns.  Note this
is only done for raw-mode as it's intended as a utility tool for
debugging decode fuzzer test outputs (for input read overflows).