Protect against RansDecRenorm overflowing input buffer.
The 4x16 decoder uses RansDecRenormSafe.  The 32-way unrolled version
only had RansDecRenorm for scalar and various forms of SIMD load.

The scalar 32-way rANS now uses RansDecRenormSafe when close to end of
input.  This is a noticable performance hit (~7% slower with clang13),
but note this only applies to systems where we have no supported SIMD
implementation available.

The SIMD variants were fixed with a different strategy.  We have a
local overflow buffer large enough for the worst case renormalisation.
We copy into it the remaining bytes of input and then do the SIMD
loads safe in the knowledge we cannot overflow.

Also correct ptr_end bounds check for rans_uncompress_O1_32x16.
It was using 8 bytes worse case for 4-way unrolling renormalisation,
but it needs to be 32-way (the NX #define).