Protect against RansDecRenorm overflowing input buffer.
The 4x16 decoder uses RansDecRenormSafe.  The 32-way unrolled version
only had RansDecRenorm for scalar and various forms of SIMD load.

The scalar 32-way rANS now uses RansDecRenormSafe when close to end of
input.  This is a noticable performance hit (~7% slower with clang13),
but note this only applies to systems where we have no supported SIMD
implementation available.

The SIMD variants were fixed with a different strategy.  We have a
local overflow buffer large enough for the worst case renormalisation.
We copy into it the remaining bytes of input and then do the SIMD
loads safe in the knowledge we cannot overflow.

Corrected ptr_end bounds check for rans_uncompress_O1_32x16.
It was using 8 bytes worse case for 4-way unrolling renormalisation,
but it needs to be 32-way (the NX #define).

Also fix the order-1 decoder overruning the compressed input stream.

This is a similar fix to the O0 case, but it already had code for
handling the overflow on output so we utilise the same bail-out
implementation for input too.

Note as with O0, SSE4 code has +72 instead of +16 as it has 8 loads
of 8 bytes each instead of 4 of 16, but those 8 loads are still
fetching 16 bytes of data (the last 8 being ignored).