Fix a fuzzing test failure in tokenise_name3.
The uncompress function didn't reset the uncompressed size (making the
subsequent assert pointless), instead trusting the recorded 'ulen'
size as real. If data uncompressed to a smaller amount, it meant parts
of the buffer were uninitialised, triggering an uninitialised memory
read.

Now this is fixed, we hit asserts, which isn't good either.  So that's
now a proper error return instead.