Also fix the order-1 decoder overruning the compressed input stream.
This is a similar fix to the O0 case, but it already had code for
handling the overflow on output so we utilise the same bail-out
implementation for input too.

Note as with O0, SSE4 code has +72 instead of +16 as it has 8 loads
of 8 bytes each instead of 4 of 16, but those 8 loads are still
fetching 16 bytes of data (the last 8 being ignored).