Added ARM to Cirrus-CI tests
Use -Werror when testing compiler flags as it gave ARM errors.
Clang doesn't return an error code when given e.g. -maxv2 if it's on a
platform that has no such option, unless -Werror is enabled.

Fix ARM Neon build when running in a subdirectory.
It's rather odd how this worked before, but for some reason the neon
code was directly #included into another source file.  That meant it
didn't work when building in a subdirectory.  It's now done in the
same way that avx2 etc work.

Also fix makedist so this source is included too.