Fix possible use of uninitialised memory in rans_uncompress_O1_4x16
Problem detected while fuzz testing with memory sanitizer.
It is possible to make an input stream that refers to unpopulated
entries in the "fb" table.  As fb was not initialised, reads
to these values were invalid.

Fix by increasing the size of the sfb_ allocation (which uses
calloc), and using the extra space for fb.  This has the side
benefit of decreasing the amount of stack used by the function.