Fix ARM Neon build when running in a subdirectory.
It's rather odd how this worked before, but for some reason the neon
code was directly #included into another source file.  That meant it
didn't work when building in a subdirectory.  It's now done in the
same way that avx2 etc work.

Also fix makedist so this source is included too.