Ensure CPU dispatch only references functions that exist
If, for example, HAVE_AVX512 is not defined, the various *_avx512
functions won't exist and cannot be referenced.  Rework the
CPU dispatch code to check the HAVE_ macros so it won't try
to use any undefined symbols.  The new version will choose
the best function that both exists, and has CPU support.

Also fixes broken indentation on 'if (do_simd) { ... }' by inverting
the check and handling the false case at the top.