Various speed improvements to rANS codecs.
The changes are a combination of likely / unlikely macros and some
code structure tweaking after extensive profiling with different
compilers on low and high-entropy data.

It's almost impossible to gain speed in all cases, and optimisations
like this usually favour one system over another.

On a modern intel system (Xeon(R) Gold 6142), a few specific compiler
versions and optimisations had issues before.  The main winners are
rans4x8 clang13 -O3 encode and gcc11 / gcc12 decodes (both 4x8 and
4x16), both often by 10-20% faster.  The biggest loss is rans4x8 clang
7.0 encode, by 5-7% slower.

On an older Intel (Xeon E5-2660), the loser is clang 10.0 rans4x8,
which is up to 5% slower.  The biggest wins are rans4x8 decode, on a
mix of compilers: gcc 7, 10, 11 and all clang.