Fix clash with overlapping buffers in 32x16 scalar O1 decoder.
The old code worked only when MAGIC2 #define was 0, as s3 and sfb
sizes were then the same.  When it's larger, setting fb to follow s3
meant the end of sfb overlapped the start of fb.

This turned up in a round-trip test fail when given a single 10MB
block of our local Novaseq qual test corpus use -o133 -c0.  It's
surprising we having seen it before.

Possibly this could be simplified now we have htscodecs_tls_alloc for
ease of reuse, but I'm keeping the historic layout for now as potentially
it has a non-TLS implementation so we still want to reduce malloc/free
overheads.