Check pthread_once() return value and print diagnostic
Usually pthread_once() will never fail. However on FreeBSD if you neglect
to link with -lpthread / -lthr, libc provides a stub that just fails with
ENOSYS, leading to a later segfault in htscodecs_tls_free() as tls is NULL.

Detect this and fail the allocation, with a log message to provide a chance
for the user to diagnose the situation.