Update hts_pack to operate in line with the spec.
Previously when too many symbols were present, we still returned
packed data but with 1 byte per symbol.  It was then up to the caller
to reject this as inappropriate use of pack.  This meant that hts_pack
was a transparent transform that always gives back something even if
it's actually just growing the data by 1 byte.

However the CRAMcodecs spec states this scenario is illegal.

The checks in rANS_static4x16pr.c were bugged for nsym==256 as there
wasn't a check of "|| out[c_meta_len] == 0", but now it's a moot point
as NULL will be returned.

Fixes #65