Remove overly aggressive buffer overflow checking in rans O1 decoder
The 4x16 O1 decoder can do compression of the O1 frequency table.
When it decoded this it was meant to ensure we have at least c_freq_sz
bytes (size of compressed table) + 16 (4 4-byte rANS states).  However
"at least" was implemented as "more than" due to a > vs >= error.

However on further inspection the code after this has an additional
check for at least 16 bytes of rANS state anyway so the freq table
check didn't need the 16 there at all.  This was good as it was
still 16 for the 32-way decoder, which again was checked, correctly,
later on.

Reported by Divon Lan