Add arith+bzip2 support to name tokeniser
Also optimises method used per data type.  For example CHAR and ALPHA
are letters/symbols and don't benefit from STRIPE mode.  Conversely
DUP, DIFF, DIGITS and DIGITS0 are always guaranteed to be 32-bit ints
so do benefit greatly.

This allows us to cut out a lot of the brute force work, offerring
faster encoding.

Benchmarks vs develop:

    Level   Old size/time           New
    1       6052123 0m1.658s        4909581 0m1.567s
    3       4924296 0m1.755s        4808368 0m1.690s (default cram3.1)
    5       4780927 0m2.859s        4768044 0m2.099s
    7       4754297 0m4.028s        4758883 0m2.279s
    9       4753731 0m4.353s        4753732 0m3.174s

    11      5998241 0m2.661s        4787219 0m2.483s
    13      4896975 0m2.994s        4703920 0m3.052s
    15      4677274 0m4.809s        4656469 0m3.915s
    17      4620982 0m7.845s        4629877 0m4.097s
    19      4620851 0m9.694s        4621212 0m6.346s

It's particularly faster at the higher compression levels, and
noticably smaller at level 1 as we only try one method but it's the
best for that type of data.

Tested on a broad mix of read names from multiple platforms and in
pos-sorted and name-sorted order.