Speed up fqz_qual_stats function
The size of the buffers is reduced 4 fold, which doesn't appear to be
detrimental to analysis while offerring better cache performance.
Also the analysis of average-qual binning strategies ("sel" parameter)
now sub-samples 1/8th of the data instead of all of it as this is
sufficient to determine where it's helpful.

It's typically a 15-20% speed up of encoding.