Refactor configure tests to work for MacOS multiarch builds
Multiarch builds compile source files twice for different
architectures and then joins the object files together using
'lipo'.  This can cause architecture-specific configure tests,
like the ones for SIMD features, to go wrong because while the
source compiles for one architecture, it may fail on the other
causing the test as a whole to incorrectly report that the feature
is not supported.

To fix this, #ifdef guards are added to both the configure checks
and the source code being built so that they won't fail on the
wrong architecture.  The configure tests are also changed so that
instead of testing to see if a particular flag works, they instead
check to see what flags are needed to build code using a
particular SIMD feature.  Due to the #ifdef guards, the tests
will work and give an empty result on the wrong architecture (the
test result will be reported as "none").  They will also do this
if no flag is needed, e.g. because other options enable the
feature already.  If the flag is needed then they will report it
and add if to CFLAGS.  Finally if the test fails completely
even though trying the correct architecture they will report
"unsupported" and disable the SIMD code.

In the multiarch case, the build on the "wrong" architecture should
now do nothing, which the "right" one will determine the flags
needed, which will be reported as the end result.

The test for NEON didn't need to try any specific flags.  As the
NEON code is selected by #ifdef __ARM_NEON already, the NEON
test has been removed and rANS_static32x16pr_neon.c is now
compiled unconditionally.

As some of the source files may now be compiled but essentially
empty due to platform #ifdef guards, some dummy symbols have
been added to avoid complaints about empty translation units.

The AX_CHECK_COMPILE_FLAG() autoconf macro is replaced with
HTS_CHECK_COMPILE_FLAGS_NEEDED() which implements the flag-checking
tests.