On AMD, don't always use AVX512 in preference to AVX2.
For now, micro-benchmarking gather instructions shows mixing 4-lanes
of AVX2 can be faster than mixing 2-lanes of AVX512.  With higher
numbers (eg 16 AVX2 registers vs 8 AVX512) things work as we'd expect
with AVX512 being competitive.  I believe this is due to long latency
times and insufficient parallelism / interleaving.

As 32-way rANS unrolling means 2 AVX512 registers of 16 states each,
or 4 AVX2 registers of 8 states each, the practical upshot is AVX2
can sometimes output AVX512 on the current 32-way rANS implementation.
(Hypothetically due to the gathers, but I'm not 100% certain it's the
only latency issue here.)

This is likely to change in future CPUs, so we may wish to revisit how
we do this selection, but for now this is a simplistic check based on
function and CPU.

Also amended the CPU detection to occur once only.  This opens the
path towards better automatic detection by measurements, but for now
this hasn't been done.

Note for CI testing purposes, when using tests/rans4x16pr -c the
switch from AVX512 to AVX2 only applies if listed in the bit field.
This -c 0x0202 is always AVX2, -c 0x0404 is always AVX512, and -c
0x0606 is selected based on encode/decode and Order 0/1.