Add a -b option to change block size for rans4x16 test.
This is purely a debugging and experimentation thing, as the block
size in the library is specified by the calling code.

It allows us to try extreme small block sizes and verify the impact of
using cpuid in a pthread_once context or always (as it was before).
With 1024 byte block size and -o4 I saw 9% faster encode and 18%
faster decode, showing the potential impact of continually checking
cpu features.  On more normal sized blocks it's irrelevant.