AVX2 rans32x16 O1 speed up.
As for AVX512 we batch up copies from the in[] array to avoid 32-way
loads every cycle.  Instead we copy 16 at a time and then do local
buffer fetching for the other 15 cycles, avoiding gathers.

This avoids cache collision problems where the buffer size is a power
of 2, which is particularly significant on AMD processors.

On "AMD EPYC 7713" with 1MB block sizes this increases -o5 -c 0x0202
speed from 156MB/s to 285MB/s. With the default (approx 1M) block size
it's 296MB/s to 301MB/s.

It doesn't impact CRAM much as our blocks are typically 10,000 or
multiples there of (eg 10,000 x 100bp quals).