Remove asserts from encode_freq1 and encode_freq_d
This has no real place in a library and causes warnings when building
with `-DNDEBUG`.

They were there to check the maths were correct in the worst case
scenario for frequency encoding.  If the maths are wrong it's trashed
memory already anyway, but it's better to die instantly than continue
leaving a potential exploit path.

Analysis of worst case encode_freq_d size:

    All freq > 8-bit is 2*256 bytes.  If alternating F[j]==0 then
    alternates 2 byte val and <0><1>, so same size.

However it's currently also not possible for all frequencies to be 128
or higher as we only permit total frequency to add up to 12-bits.  I
think the most we can get is 31 2-byte freqs and 225 1-byte freq, but
with alternating 0s that would be 31*2 + 112*1 + 113*2 (0s) = 400
bytes.

Analysis of worst case encode_alphabet:

    With no RLE, it's 257 bytes (256 + nul termination)
    With RLE it's <= 257.

Analysis of worst case encode_freq1:

    <0> + [encode_alphabet] + 256*[encode_freq_d]
    ie 1 + 257 + 256*256*2  (or even 1+257+256*400)
     < 257*256*2

The assert was 257*257*3, along with the size in
rans_compress_bound_4x16, which is harmlessly larger than needed.
(This size likely came from the old 4x8 rANS codec which had a
different O1 encoding layout.)

Note a similar pointless assert exists in rans_compress_O1, but I
haven't reevaluated the calculations there and it doesn't trigger the
same warning as tab_size is used elsewhere, so I'm ignoring that one
for now.