Add bounds checking to range coder.
This is optional and enabled by a new function RC_SetOutputEnd.
It already existed for decoding as we can't validate the input data,
but the encoder had an explicit assumption that the caller new what it
was doing.  We now double check.  The performance hit for encode
varies by compiler, but at worse I found 9% slow down and at best a
small speed up.  Encode rates in MB/s:

          master        checked
          o0   o1       o0   o1
gcc7      44.2 39.2     44.7 37.5
gcc13     48.4 40.3     43.3 39.8
clang10   38.9 35.5     39.7 32.2
clang16   42.0 35.5     42.0 37.6

Note however it's unlikely arith_dynamic would have been able to
trigger any issues as arith_compress_bound was already being used to
validate the maximum output buffer size. However there was a lack of
validation of errors even for the decoder as it was a void function.
It now returns an error code so arith_decompress_* can return NULL,
making things fail sooner.