Remove divide by zero in rANS 4x8.
When given 0 bytes of input, we get a floating point exception in the
code attempting to renormalise the frequencies.

The test harness can't trigger this case as 0 is EOF in the read loop,
but that has been changed so a null file still calls the compression.
Also tested the other entropy encoders, but they already coped fine.

Bug reported by Shubham Chandak