Fix mem bounds checking in rans_uncompress_O0_32x16_avx2.
The location of the sp > cp_end check was previously moved to the
later position used in the gcc code-path.  However this was inadequate
for actually preventing buffer read-overruns.  Note this is also true
in the master branch.

It's now back higher again, closer to the old __clang__ ifdef
position, but not quite.  This position appears to resolve the big
speed loss we had with clang10 with poor placement, while not harming
the other compilers, and obviously fixing the buffer bounds check.