Correct a buffer read-overflow (max 8 bytes) in AVX2 rans decode
The check for 'sp > cp_end' either needs to be before all the
mm256_loads or after all of them (ie start or end of loop), but the
previous placement was just before the last "sp += popcnt..." line.

On a block of 135 zero bytes this was giving read-buffer overflow
errors. (Reported by Rob Davies)