Add explicit -mpopcnt to avx2 and avx512 builds.
Zig adds native support for the CPU being built on.  So if we support
popcnt then we don't need an explicit -mpopcnt.  However adding any -m
options appears to disable that native support, so -mavx2 doesn't
enable -mpopcnt, yet our auto-detection previously claimed it was
necessary.  This means building on a CPU with popcnt but no avx2 will
fail to build (with zig) when targetting the avx2 platform.

Therefore the only reliable way to check is the combinations actively
in use.  So if we want to compile code using popcnt + avx2 then we
have to test -mpopcnt -mavx2 together.

It's problematic and tricky to manage in autoconf.

Fixes #109