Remove undefined behaviour in transpose_and_copy.
There was a -DUBSAN guard for use with undefined behaviour sanitizers,
but actually the only code change that's necessary is to change the
explicit writes to a memcpy.  This fixes unoptimised zig-cc builds
which use -fsanitize=undefined by default.

I've tested gcc7, gcc13, clang10, clang16, zig cc, icc and icl.  All
seem fine with this change.

NB: Zig cc is the only on to vectorise the most naive implementation,
and it does so very well, basically creating an SSE4 version of the
loop. That is the even fastest implementation for zig cc.