Improve the UUID4 detection.
We already had code for spotting this in search_trie, so improved that
a little and use it in encode_name instead of having a second scan.

Also improve the compression of mixed data sets.  This still isn't
optimal as we'd need to start separating the name classes and adding
NOP tokens, but it's often a 10-20% compression improvement.