Protection from attempting to do SIMD rans encoding on small data
We can't just change do_simd as we may then have a combination of SIMD
(pack or rle-meta) and non-SIMD (rle-meta or rle-data).  So instead we
remove the ability to do SIMD.

This only happens when we have large blocks of data that enable SIMD
mode which then turn out to be heavily compressible after PACK or RLE
to yield a small sub-stream which isn't SIMD compressible.