Add dedicated UUID4 (ONT) recognition to the name tokeniser (PR #134)
Improved existing code for spotting UUIDs in search_trie, and updated
encode_name to store them when found.

Also improves the compression of mixed data sets.  This still isn't
optimal as we'd need to start separating the name classes and adding
NOP tokens, but it's often a 10-20% compression improvement.

We still need to do more work on optimising name compression in a
general manner, especially on mixed data sets, but this is resolves a
simple case while having no impact on any other name formats.