Add cpuid checks for XSAVE, OSXSAVE and AVX
AVX2 and AVX512F depend on AVX, XSAVE and OSXSAVE being present.
If they are not, AVX2/AVX512F instructions may be blocked even
though cpuid() reports them to be available.  Spotted when the
AVX512 rANS codec crashed with illegal instructions on a virtual
host that claimed to have AVX512F, but did not have OSXSAVE.