Check XCR0 register before using AVX2 / AVX512 instructions
Along with checking CPUID, it's necessary to look in the XCR0
register to check that AVX, AVX2 and AVX512 instructions can
be used.  (The operating system can write to this register to
selectively enable or disable these features).  See the Intel 64
and IA-32 Architectures Software Developer’s Manual Vol. 1
sections 13.2 and 13.3 for details.

XCR0 is read using the XGETBV instruction.  While there is an
intrinsic for this, using it requires specific compiler options
that we may not want to use for htscodecs/rANS_static4x16pr.c
compilation.  The intrinsic also didn't work until gcc 9.

As some binutils still in use may not know about XGETBV, the
instruction is written as a byte stream in the inline assembly.